pyproject.toml
[project]
name = "cvmaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Sangwon Lee", email = "lsw91.main@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.29.5",
    "jinja2>=3.1.6",
    "markdown-table-generator>=1.1.0",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "scholarly",
    "pyppeteer>=1.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
scholarly = { git = "https://github.com/scholarly-python-package/scholarly.git" }
